+++ core/src/main/java/hud/EventosBoton.java
@@ -5,27 +5,27 @@ import com.principal.AntsArmageddon;
 import screens.GameScreen;
 import screens.MenuScreen;
 import screens.OpcionesScreen;
+import utils.Utiles;
 
 public class EventosBoton {
 
-    public static Runnable salirDelJuego(){
-        return() -> Gdx.app.exit();
-    }
-
     public static Runnable irMenuOpciones(AntsArmageddon juego){
-        return() -> juego.setScreen(new OpcionesScreen(juego));
+        return() -> juego.setScreen(new OpcionesScreen(juego, juego.getAssetManager()));
     }
 
     public static Runnable salirMenuOpciones(AntsArmageddon juego){
-        return() -> juego.setScreen(new MenuScreen(juego));
+        return() -> juego.setScreen(new MenuScreen(juego, juego.getAssetManager()));
     }
 
     public static Runnable irJuego(AntsArmageddon juego){
-        return() -> juego.setScreen(new GameScreen(juego));
+        return() -> juego.setScreen(new GameScreen(juego, juego.getAssetManager()));
     }
 
-    public static Runnable salirJuego(AntsArmageddon juego){
-        return() -> juego.setScreen(new MenuScreen(juego));
+    public static Runnable salirJuego(){
+        return() -> Gdx.app.exit();
     }
 
+    public static Runnable descomponerAtlas(){
+        return() -> Utiles.descomponerAtlas("botones/pack.atlas", "atlasDescompuestos/");
+    }
 }