+++ core/src/main/java/screens/MenuScreen.java
@@ -2,6 +2,8 @@ package screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.assets.AssetManager;
+import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.Texture;
@@ -9,6 +11,7 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
+import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.principal.AntsArmageddon;
@@ -19,6 +22,7 @@ import utils.Constantes;
 public class MenuScreen implements Screen {
 
     private final AntsArmageddon juego;
+    private AssetManager assetManager;
 
     private Stage escenario;
     private FitViewport viewport;
@@ -27,18 +31,20 @@ public class MenuScreen implements Screen {
     private Sprite spriteFondo;
     private SpriteBatch batch;
 
-    public MenuScreen(AntsArmageddon juego){
+    public MenuScreen(AntsArmageddon juego, AssetManager assetManager){
+        this.assetManager = assetManager;
         this.juego = juego;
+
     }
 
     @Override
     public void show() {
         camara = new OrthographicCamera();
-        viewport = new FitViewport(Constantes.ANCHO, Constantes.ALTO, camara);
+        viewport = new FitViewport(Constantes.RESOLUCION_ANCHO, Constantes.RESOLUCION_ALTO, camara);
         escenario = new Stage(viewport);
         Gdx.input.setInputProcessor(escenario);
 
-        texturaFondo = new Texture(Gdx.files.internal("fondoPantalla.png"));
+        texturaFondo = assetManager.get(Constantes.FONDO_JUEGO, Texture.class);
         spriteFondo = new Sprite(texturaFondo);
 
         spriteFondo.setSize(viewport.getWorldWidth(), viewport.getWorldHeight());
@@ -62,14 +68,16 @@ public class MenuScreen implements Screen {
 
     }
 
-    public void construirMenu(){
-        ImageButton jugar = FabricaBotones.crearBoton("jugar.png");
-        ImageButton opciones = FabricaBotones.crearBoton("opciones.png");
-        ImageButton salir = FabricaBotones.crearBoton("salir.png");
+    public void construirMenu() {
+        ImageButton jugar = FabricaBotones.crearBoton(assetManager, "jugar_up", "jugar_over", "jugar_down", EventosBoton.irJuego(juego));
+        ImageButton opciones = FabricaBotones.crearBoton(assetManager, "opciones_up", "opciones_over", "opciones_down", EventosBoton.irMenuOpciones(juego));
+        ImageButton salir = FabricaBotones.crearBoton(assetManager, "salir_up", "salir_over", "salir_down", EventosBoton.salirJuego());
 
-        FabricaBotones.agregarEventos(jugar, EventosBoton.irJuego(juego));
-        FabricaBotones.agregarEventos(opciones, EventosBoton.irMenuOpciones(juego));
-        FabricaBotones.agregarEventos(salir, EventosBoton.salirJuego(juego));
+        Sound sonidoClick = assetManager.get(Constantes.SONIDO_BOTONES, Sound.class);
+
+        FabricaBotones.agregarSonido(jugar, sonidoClick);
+        FabricaBotones.agregarSonido(opciones, sonidoClick);
+        FabricaBotones.agregarSonido(salir, sonidoClick);
 
         Table table = new Table();
         table.setFillParent(true);
@@ -83,11 +91,12 @@ public class MenuScreen implements Screen {
 
 
 
+
     @Override
     public void resize(int ancho, int alto) { viewport.update(ancho, alto, true); }
 
     @Override
-    public void dispose() { escenario.dispose(); texturaFondo.dispose(); }
+    public void dispose() { escenario.dispose(); batch.dispose(); }
 
     @Override
     public void pause() { }