+++ core/src/main/java/utils/Constantes.java
@@ -2,6 +2,11 @@ package utils;
 
 public class Constantes {
 
-    public static final int ANCHO = 1280;
-    public static final int ALTO = 720;
+    public static final String ATLAS_BOTONES = "botones/botones.atlas";
+    public static final String SONIDO_BOTONES = "sonido_click.mp3";
+    public static final String FONDO_JUEGO = "fondoPantalla.png";
+    public static final String FONDO_GAMEPLAY = "Mapa.png";
+
+    public static final int RESOLUCION_ANCHO = 1280;
+    public static final int RESOLUCION_ALTO = 720;
 }